# openapi/sesn.v1.yaml
openapi: 3.1.0
info:
  title: SES Navigator API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /contracts:
    post:
      operationId: createContract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContractInput' }
      responses:
        '201':
          description: Draft contract created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      operationId: listContracts
      parameters:
        - {
            name: q,
            in: query,
            schema: { type: string, minLength: 1, maxLength: 100 },
          }
        - {
            name: status,
            in: query,
            schema: { $ref: '#/components/schemas/ContractStatus' },
          }
        - { name: cursor, in: query, schema: { type: string } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 200, default: 50 },
          }
      responses:
        '200':
          description: Safe contract summary list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContractList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /contracts/{id}:
    put:
      operationId: updateContract
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: If-Match,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContractInput' }
      responses:
        '200':
          description: Draft contract updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    get:
      operationId: getContract
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        '200':
          description: Restricted reviewed contract detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /contracts/{id}/status:
    post:
      operationId: transitionContractStatus
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: If-Match,
            in: header,
            required: true,
            schema: { type: string },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: '#/components/schemas/ContractStatusTransitionInput' }
      responses:
        '200':
          description: Contract approval status transitioned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ContractType:
      type: string
      enum: [ses, dispatch, subcontract, quasi_mandate, fixed_price, other]
    ContractStatus:
      type: string
      enum: [draft, review, active, suspended, expired, terminated, cancelled]
    ContractSummary:
      type: object
      required:
        [
          id,
          contractNo,
          projectId,
          proposalId,
          engineerId,
          contractType,
          status,
          title,
          startDate,
          endDate,
          autoRenew,
          currency,
          updatedAt,
          rowVersion,
        ]
      properties:
        id: { type: string, format: uuid }
        contractNo: { type: string }
        projectId: { type: [string, 'null'], format: uuid }
        proposalId: { type: [string, 'null'], format: uuid }
        engineerId: { type: [string, 'null'], format: uuid }
        contractType: { $ref: '#/components/schemas/ContractType' }
        status: { $ref: '#/components/schemas/ContractStatus' }
        title: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: [string, 'null'], format: date }
        autoRenew: { type: boolean }
        currency: { type: string, minLength: 3, maxLength: 3 }
        updatedAt: { type: string, format: date-time }
        rowVersion: { type: integer, minimum: 1 }
    Contract:
      allOf:
        - { $ref: '#/components/schemas/ContractSummary' }
        - type: object
          required:
            [
              monthlyAmount,
              hourlyAmount,
              settlementLowerHours,
              settlementUpperHours,
              paymentTerms,
              notes,
              parties,
              versions,
              workLogs,
              approval,
            ]
          properties:
            monthlyAmount: { type: [number, 'null'], minimum: 0 }
            hourlyAmount: { type: [number, 'null'], minimum: 0 }
            settlementLowerHours: { type: [number, 'null'], minimum: 0 }
            settlementUpperHours: { type: [number, 'null'], minimum: 0 }
            paymentTerms: { type: [string, 'null'] }
            notes: { type: [string, 'null'] }
            parties:
              type: array
              items: { $ref: '#/components/schemas/ContractParty' }
            versions:
              type: array
              items: { $ref: '#/components/schemas/ContractVersion' }
            workLogs:
              type: array
              items: { $ref: '#/components/schemas/ContractWorkLog' }
            approval:
              oneOf:
                - { $ref: '#/components/schemas/ContractApproval' }
                - { type: 'null' }
    ContractParty:
      type: object
      additionalProperties: false
      required: [id, companyId, contactId, partyRole, billingRole, isPrimary]
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        contactId: { type: [string, 'null'], format: uuid }
        partyRole:
          {
            type: string,
            enum:
              [
                customer,
                supplier,
                employer,
                end_client,
                prime_contractor,
                subcontractor,
                other,
              ],
          }
        billingRole:
          { type: [string, 'null'], enum: [bill_to, pay_to, none, null] }
        isPrimary: { type: boolean }
    ContractPartyInput:
      type: object
      additionalProperties: false
      required: [companyId, contactId, partyRole, billingRole, isPrimary]
      properties:
        companyId: { type: string, format: uuid }
        contactId: { type: [string, 'null'], format: uuid }
        partyRole:
          {
            type: string,
            enum:
              [
                customer,
                supplier,
                employer,
                end_client,
                prime_contractor,
                subcontractor,
                other,
              ],
          }
        billingRole:
          { type: [string, 'null'], enum: [bill_to, pay_to, none, null] }
        isPrimary: { type: boolean }
    ContractApproval:
      type: object
      additionalProperties: false
      required:
        [
          id,
          status,
          requestedAt,
          completedAt,
          requestNote,
          decisionNote,
          rowVersion,
        ]
      properties:
        id: { type: string, format: uuid }
        status:
          {
            type: string,
            enum: [draft, pending, approved, rejected, cancelled, expired],
          }
        requestedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        requestNote: { type: [string, 'null'], maxLength: 1000 }
        decisionNote: { type: [string, 'null'], maxLength: 1000 }
        rowVersion: { type: integer, minimum: 1 }
    ContractInput:
      type: object
      additionalProperties: false
      required:
        [
          contractNo,
          projectId,
          proposalId,
          engineerId,
          contractType,
          title,
          startDate,
          endDate,
          autoRenew,
          currency,
          monthlyAmount,
          hourlyAmount,
          settlementLowerHours,
          settlementUpperHours,
          paymentTerms,
          notes,
          parties,
          changeSummary,
        ]
      properties:
        contractNo: { type: string, minLength: 1, maxLength: 32 }
        projectId: { type: [string, 'null'], format: uuid }
        proposalId: { type: [string, 'null'], format: uuid }
        engineerId: { type: [string, 'null'], format: uuid }
        contractType: { $ref: '#/components/schemas/ContractType' }
        title: { type: string, minLength: 1, maxLength: 300 }
        startDate: { type: string, format: date }
        endDate: { type: [string, 'null'], format: date }
        autoRenew: { type: boolean }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        monthlyAmount: { type: [number, 'null'], minimum: 0 }
        hourlyAmount: { type: [number, 'null'], minimum: 0 }
        settlementLowerHours: { type: [number, 'null'], minimum: 0 }
        settlementUpperHours: { type: [number, 'null'], minimum: 0 }
        paymentTerms: { type: [string, 'null'], maxLength: 1000 }
        notes: { type: [string, 'null'], maxLength: 5000 }
        parties:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/ContractPartyInput' }
        changeSummary: { type: [string, 'null'], maxLength: 1000 }
    ContractStatusTransitionInput:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { type: string, enum: [draft, review, active] }
        reason: { type: [string, 'null'], maxLength: 1000 }
    ContractVersion:
      type: object
      additionalProperties: false
      required:
        [
          id,
          versionNo,
          effectiveFrom,
          effectiveTo,
          changeSummary,
          approvedAt,
          createdAt,
        ]
      properties:
        id: { type: string, format: uuid }
        versionNo: { type: integer, minimum: 1 }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: [string, 'null'], format: date }
        changeSummary: { type: [string, 'null'] }
        approvedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    ContractWorkLog:
      type: object
      additionalProperties: false
      required:
        [
          id,
          engineerId,
          workMonth,
          status,
          scheduledDays,
          actualDays,
          scheduledHours,
          actualHours,
          overtimeHours,
          absenceHours,
          customerApprovedAt,
          updatedAt,
          rowVersion,
        ]
      properties:
        id: { type: string, format: uuid }
        engineerId: { type: string, format: uuid }
        workMonth: { type: string, format: date }
        status:
          { type: string, enum: [draft, submitted, approved, rejected, locked] }
        scheduledDays: { type: [number, 'null'], minimum: 0 }
        actualDays: { type: [number, 'null'], minimum: 0 }
        scheduledHours: { type: [number, 'null'], minimum: 0 }
        actualHours: { type: [number, 'null'], minimum: 0 }
        overtimeHours: { type: number, minimum: 0 }
        absenceHours: { type: number, minimum: 0 }
        customerApprovedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: string, format: date-time }
        rowVersion: { type: integer, minimum: 1 }
    ContractList:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          {
            type: array,
            items: { $ref: '#/components/schemas/ContractSummary' },
          }
        page:
          type: object
          required: [limit, nextCursor]
          properties:
            limit: { type: integer }
            nextCursor: { type: ['string', 'null'] }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
  responses:
    BadRequest:
      {
        description: Invalid request,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    Conflict:
      {
        description: Optimistic locking conflict,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    PreconditionRequired:
      {
        description: If-Match is required,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    Unauthorized:
      {
        description: Authentication required,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    Forbidden:
      {
        description: Permission denied,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }
    NotFound:
      {
        description: Not found or not visible,
        content:
          {
            application/json:
              { schema: { $ref: '#/components/schemas/ApiError' } },
          },
      }